% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_NH.R
\name{acceptance_probability}
\alias{acceptance_probability}
\title{To compute acceptance probability}
\usage{
acceptance_probability(old_cost, new_cost, Temp)
}
\arguments{
\item{old_cost}{Old solution cost}

\item{new_cost}{New solution cost}

\item{Temp}{Temperature}
}
\description{
To compute acceptance probability
}
