% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A-n32-k5-locations-data.R
\docType{data}
\name{An32k5locations}
\alias{An32k5locations}
\title{Sample dataset for VRP (32 nodes, including depot)}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\source{
\href{http://neo.lcc.uma.es/vrp/vrp-instances/capacitated-vrp-instances/}{QTL Archive}
}
\usage{
data(An32k5locations)
}
\description{
Locations data for 31 customer nodes
}
\keyword{datasets}
