% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckConstraints.R
\name{CheckConstraints}
\alias{CheckConstraints}
\title{Checks whether the constraints satisfied or not}
\usage{
CheckConstraints(route, common_node, new_node, demand, Vehicle_Capacity,
  Constraints = c("Capacity"))
}
\arguments{
\item{route}{-- Sequence of nodes (Ex: 5 4 3 8)}

\item{common_node}{-- One of the existing node of the route and it shouldn't be an interior of the route (Interior of route: nodes, which have both left and right edge, ex: 3, 4 in above route) (Ex: 5)}

\item{new_node}{-- New node to be included to the route (Ex: 9)}

\item{Constraints}{-- List of constraints to check. Currently implemented only vehicle capacity constraint}
}
\description{
Checks whether the constraints satisfied or not
}
\examples{
route <- c(5, 4, 3, 8)
common_node <- 5
new_node <- 9
Vehicle_Capacity <- 100
data(An32k5demand)
demand <- An32k5demand
const <- c("Capacity")
CheckConstraints(route, common_node, new_node, demand, Vehicle_Capacity, Constraints = const)
}
