% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connect_Edges.R
\name{Connect_Edges}
\alias{Connect_Edges}
\title{Connect new edges with the existing route}
\usage{
Connect_Edges(route, common_node, new_node)
}
\arguments{
\item{route}{-- Sequence of nodes (Ex: 5 4 3 8)}

\item{common_node}{-- One of the existing node of the route (Ex: 5)}

\item{new_node}{-- New node to be included to the route (Ex: 9)}
}
\description{
Connect new edges with the existing route
}
\examples{
route <- c(5, 4, 3, 8)
common_node <- 5
new_node <- 9
Connect_Edges(route, common_node, new_node)
}
