% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crossover.R
\name{crossover}
\alias{crossover}
\title{Performs Order Crossover (OX) between two chromozomes}
\usage{
crossover(parents)
}
\arguments{
\item{parents}{Selected parents}
}
\description{
Performs Order Crossover (OX) between two chromozomes
}
\examples{
n <- 10 # Number of nodes
parents <- rbind(sample(n, n), sample(n, n))
crossover(parents)
}
