% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParallelRouting.R
\name{CW_Parallel_VRP}
\alias{CW_Parallel_VRP}
\title{Performs Clarke-Wright Savings algorithm in Parallel}
\usage{
CW_Parallel_VRP(Sort_Edge, nnodes, demand, Vehicle_Capacity, routes = list(),
  Served_nodes = c(), Filled_Routes = c(), logfile = TRUE,
  Constraints = c("Capacity"))
}
\arguments{
\item{Sort_Edge}{-- Edges sorted in a decreasing order of their savings}

\item{nnodes}{-- Number of customer nodes}
}
\description{
Performs Clarke-Wright Savings algorithm in Parallel
}
