% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequentialRouting.R
\name{CW_Sequential_VRP}
\alias{CW_Sequential_VRP}
\title{Performs Clarke-Wright Savings algorithm in Sequential}
\usage{
CW_Sequential_VRP(Sort_Edge, nnodes, demand, Vehicle_Capacity,
  routes = list(), Served_nodes = c(), Filled_Routes = c(),
  logfile = TRUE, Constraints = c("Capacity"))
}
\arguments{
\item{Sort_Edge}{-- Edges sorted in a decreasing order with respect to savings}

\item{nnodes}{-- Number of customer nodes (Excluding depot)}
}
\description{
Performs Clarke-Wright Savings algorithm in Sequential
}
