% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistMat.R
\name{DistMat}
\alias{DistMat}
\title{Coomputes distance matrix}
\usage{
DistMat(locations, method = "euclidean")
}
\arguments{
\item{locations}{-- [ID, X, Y] Node id and X, Y co-ordinates/ Long, Lat}

\item{method}{-- Metric to calculate distnace between nodes. Feasible methods for X-Y co-ordinates c("euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski"); Feasible methods for Long-Lat c(distCosine, distHaversine). Default: "euclidean". Note: Please make sure that, method should be a character only for X-Y co-ordinates not for Long-Lat}
}
\description{
Coomputes distance matrix
}
\examples{
data(An32k5locations)
DistMat(An32k5locations)
}
