% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_Index_Common_Edge.R
\name{Find_Ind_Common_Edge}
\alias{Find_Ind_Common_Edge}
\title{To find an index of the sorted edges, in which a common node between routes and existing routes exists. (Only for Sequential approach, internal purpose)}
\usage{
Find_Ind_Common_Edge(Sort_Edge, Served_nodes)
}
\arguments{
\item{Sort_Edge}{-- Edges sorted in a decreasing order of their savings}

\item{Served_nodes}{-- List of nodes already serviced}
}
\description{
To find an index of the sorted edges, in which a common node between routes and existing routes exists. (Only for Sequential approach, internal purpose)
}
