% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness.R
\name{fitness}
\alias{fitness}
\title{Computes fitness of a solution. It handles vehicle capacity.}
\usage{
fitness(sol, vc, demand, DMat, route = FALSE)
}
\arguments{
\item{sol}{Solution vector}

\item{vc}{Vehicle capacity}

\item{route}{Logical variable. If TRUE,}
}
\description{
Computes fitness of a solution. It handles vehicle capacity.
}
\examples{
locations <- An32k5locations
demand <- An32k5demand
sol <- sample(nrow(locations)-1, nrow(locations)-1)
vc <- 100
DMat <- as.matrix(dist(locations[, -1], upper = TRUE, diag = TRUE))
row.names(DMat) <- c(0, 1:(nrow(demand)-1))
colnames(DMat) <- c(0, 1:(nrow(demand)-1))
fitness(sol, vc, demand, DMat, route = TRUE)
}
