% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneticAlgoVRP.R
\name{GA_VRP}
\alias{GA_VRP}
\title{Genetic Algorithm structure to solve VRP}
\usage{
GA_VRP(maxitr, runs, cp, mp, fitness = fitness, selection = selection_RS,
  crossover = crossover, mutation = mutation, init_pop = NULL,
  num_nodes = NULL, popSize = NULL)
}
\arguments{
\item{maxitr}{Maximum number of iterations}

\item{runs}{If there is no improvement in the fitness value in the consecutive 'runs' iterations then algorithm will terminate}

\item{cp}{Crossover probability}

\item{mp}{Mutation probability}

\item{fitness}{Fitness function}

\item{selection}{Fuction for selection operator, default selection_RS (Selection with rank space but not with diversity)}

\item{crossover}{Function for crossover operator}

\item{mutation}{Function for mutation operator}

\item{init_pop}{Initial population for GA}

\item{num_nodes}{Number of customer nodes}

\item{popSize}{Number of chromozomes in population}
}
\description{
Genetic Algorithm structure to solve VRP
}
