% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_NH.R
\name{generate_neighbor}
\alias{generate_neighbor}
\title{Function to generate neighbor solution. It's an internal function to SA_VRP.}
\usage{
generate_neighbor(solution)
}
\arguments{
\item{solution}{solution vector}
}
\description{
Function to generate neighbor solution. It's an internal function to SA_VRP.
}
\examples{
solution <- sample(10, 10)
generate_neighbor(solution)
}
