% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isInterior.R
\name{isInterior}
\alias{isInterior}
\title{To find whether a node is interior of the route or not}
\usage{
isInterior(route, node)
}
\arguments{
\item{route}{-- Sequence of nodes (Ex: 5 4 3 8)}

\item{node}{-- One of the existing node of the route (Ex: 5)}
}
\description{
To find whether a node is interior of the route or not
}
\examples{
route <- c(5, 4, 3 ,8)
node <- 5
isInterior(c(5, 4, 3, 8), 5) # FALSE
isInterior(c(5, 4, 3, 8), 4) # TRUE
}
