% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mutation.R
\name{mutation}
\alias{mutation}
\title{Mutates parent chromozome}
\usage{
mutation(parent)
}
\arguments{
\item{parent}{One parent chromozome}
}
\description{
Mutates parent chromozome
}
\examples{
n <- 10 # Number of nodes
parent <- sample(n, n)
mutation(parent)
}
