% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SA_VRP.R
\name{SA_VRP}
\alias{SA_VRP}
\title{To find near optimal solutions using Simmulated Annealing.}
\usage{
SA_VRP(fitness, vc, demand, locations, method = "euclidean", popSize = NULL,
  nrun = 25)
}
\arguments{
\item{fitness}{Fitness function}

\item{vc}{Vehicle capacity}

\item{demand}{Demand at each node}

\item{method}{-- Metric to calculate distnace between nodes. Feasible methods for X-Y co-ordinates c("euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski"); Feasible methods for Long-Lat c(distCosine, distHaversine). Default: "euclidean". Note: Please make sure that, method should be character for X-Y co-ordinates and not for Long-Lat.}

\item{popSize}{Population Size}

\item{nrun}{Number of runs}

\item{location}{Location of each node}
}
\description{
To find near optimal solutions using Simmulated Annealing.
}
\examples{
popSize <- 100
num_nodes <- 31
fitness <- fitness
vc <- 100
Demand <- An32k5demand
Locations <- An32k5locations
DMat <- DistMat(locations)
row.names(DMat) <- c(0, 1:(nrow(demand)-1))
colnames(DMat) <- c(0, 1:(nrow(demand)-1))
SA_VRP(fitness, vc, demand, locations, popSize = 100)
SA_VRP(fitness, vc, demand, locations)
}
