% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SavingMat.R
\name{SavingMat}
\alias{SavingMat}
\title{To compute savings matrix}
\usage{
SavingMat(DMat, depot = 1)
}
\arguments{
\item{DMat}{-- Distance matrix, assuming first row and column represents distance from depot
depot = 1 by default (fixed)}
}
\description{
To compute savings matrix
}
\examples{
data(An32k5locations)
DMat <- DistMat(An32k5locations)
row.names(DMat) <- An32k5locations[, 1]
colnames(DMat) <- An32k5locations[, 1]
SavingMat(DMat)
}
