% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionWithDiversityInternal.R
\name{selection_diversity}
\alias{selection_diversity}
\title{Rank space selection operator with diversity, internal purpose (look at selection for more details)}
\usage{
selection_diversity(prob, edge_dist, diversity, sel = NULL, pc, p, i = 1)
}
\arguments{
\item{prob}{Selection probability values, it is a function of fitness values}

\item{edge_dist}{Distance between two connected edges}

\item{sel}{Index of selected units of population}

\item{pc}{Crossover probability}

\item{p}{Weight for diversification}

\item{i}{Counter, internal purpose}
}
\description{
Rank space selection operator with diversity, internal purpose (look at selection for more details)
}
