% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectionWithDiversity.R
\name{selection_RS_Diversity}
\alias{selection_RS_Diversity}
\title{Rank space selection operator with diversity}
\usage{
selection_RS_Diversity(fit, edge_dist, pc = 0.6, p = 0.2)
}
\arguments{
\item{fit}{Fitness values of population}

\item{edge_dist}{Distance between two connected nodes}

\item{pc}{Crossover probability}

\item{p}{Weight for diversification}
}
\description{
Rank space selection operator with diversity
}
\examples{

}
