% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sorted_Edges.R
\name{Sorted_Edges}
\alias{Sorted_Edges}
\title{To sort edges in a decreasing order of their savings}
\usage{
Sorted_Edges(DMat)
}
\arguments{
\item{DMat}{-- Distance matrix}
}
\description{
To sort edges in a decreasing order of their savings
}
\examples{
data(An32k5locations)
DMat <- DistMat(An32k5locations)
row.names(DMat) <- An32k5locations[, 1]
colnames(DMat) <- An32k5locations[, 1]
Sorted_Edges(DMat)
}
