% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Cost.R
\name{Total_Cost}
\alias{Total_Cost}
\title{To compute total cost of the routes}
\usage{
Total_Cost(result, DMat)
}
\arguments{
\item{result}{-- List of routes}
}
\description{
To compute total cost of the routes
}
\examples{
data(An32k5locations)
locations <- An32k5locations
DMat <- DistMat(locations)
row.names(DMat) <- locations[, 1]
colnames(DMat) <- locations[, 1]
data(An32k5demand)
demand <- An32k5demand
Vehicle_Capacity <- 100
result <- CW_VRP(demand, DMat = DMat, Vehicle_Capacity = Vehicle_Capacity)
Total_Cost(result, DMat)
}
